fix(answer) : 트랜잭션 분리
서현님 피드백 반영
