+++ src/main/java/com/gamyeon/answer/application/AnswerAnalysisProperties.java
@@ -12,4 +12,5 @@ public class AnswerAnalysisProperties {
   private int dispatchBatchSize = 10;
   private int maxRetryCount = 5;
   private long schedulerFixedDelayMs = 60_000L;
+  private long sendingRecoveryTimeoutMs = 300_000L;
 }
